loungeHistory only did not work due to dirty db removed delay in setup lint clean up
