+++ lib/ui/screens/my_account/widgets/add_account_options_widget.dart
@@ -1,8 +1,4 @@
-import 'package:app_2i2i/infrastructure/data_access_layer/accounts/local_account.dart';
 import 'package:app_2i2i/infrastructure/routes/app_routes.dart';
-import 'package:app_2i2i/ui/commons/custom_dialogs.dart';
-
-import 'package:app_2i2i/ui/screens/my_account/create_local_account.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -17,7 +13,6 @@ import '../../../../infrastructure/data_access_layer/accounts/walletconnect_acco
 import '../../../../infrastructure/data_access_layer/services/logging.dart';
 import '../../../../infrastructure/providers/all_providers.dart';
 import '../../../../infrastructure/providers/my_account_provider/my_account_page_view_model.dart';
-import 'keys_widget.dart';
 import 'qr_image_widget.dart';
 
 class AddAccountOptionsWidgets extends ConsumerStatefulWidget {
@@ -123,8 +118,8 @@ class _AddAccountOptionsWidgetsState
           ),
           ListTile(
             onTap: () async {
-                Navigator.of(context).maybePop();
-                context.pushNamed(Routes.createLocalAccount.nameFromPath());
+              Navigator.of(context).maybePop();
+              context.pushNamed(Routes.createLocalAccount.nameFromPath());
             },
             leading: Container(
               height: 50,
@@ -156,21 +151,9 @@ class _AddAccountOptionsWidgetsState
       ),
     );
   }
-  var val;
-  onClickVerify(Map value){
-    Navigator.of(context,rootNavigator: true).pop();
-    val = value;
-    print('value before $value');
-    Future.delayed(Duration(seconds: 2)).then((value) {
-      print('value after $val');
-      Navigator.of(buildContext).pushNamed(
-        Routes.verifyPerhaps,
-        arguments:value,
-      );
-    });
-  }
 
-  Future _createSession(MyAccountPageViewModel myAccountPageViewModel, AccountService accountService) async {
+  Future _createSession(MyAccountPageViewModel myAccountPageViewModel,
+      AccountService accountService) async {
     final account = WalletConnectAccount.fromNewConnector(
       accountService: accountService,
     );
@@ -185,7 +168,7 @@ class _AddAccountOptionsWidgetsState
       await account.setMainAccount();
       _displayUri = '';
       if (isDialogOpen) {
-        Navigator.of(context,rootNavigator: true).pop();
+        Navigator.of(context, rootNavigator: true).pop();
         Navigator.of(context).pop();
       }
     } else {
