+++ lib/ui/screens/search/widgtes/user_info_tile.dart
@@ -11,7 +11,6 @@ import '../../../../infrastructure/providers/all_providers.dart';
 import '../../../../infrastructure/routes/app_routes.dart';
 import '../../../commons/custom.dart';
 import '../../../commons/custom_profile_image_view.dart';
-import '../../user_info/user_info_page.dart';
 
 class UserInfoTile extends ConsumerWidget {
   final Hangout hangout;
