+++ lib/ui/screens/top/widgets/top_durations_page.dart
@@ -2,9 +2,7 @@ import 'package:app_2i2i/infrastructure/commons/utils.dart';
 import 'package:app_2i2i/infrastructure/models/meeting_model.dart';
 import 'package:app_2i2i/infrastructure/providers/all_providers.dart';
 import 'package:app_2i2i/infrastructure/routes/app_routes.dart';
-
 import 'package:app_2i2i/ui/screens/home/wait_page.dart';
-import 'package:app_2i2i/ui/screens/user_info/user_info_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
